Replace place-only on-demand geocode smoke with a multi-type harness.
